+++ src/compas_cadwork/elements/__init__.py
@@ -3,8 +3,10 @@ from .dimensional_element import DimensionalElement
 from .element import Element
 from .element_type import ElementType
 from .factory import AnyElement
+from .floor import Floor
 from .oriented_element import OrientedElement
 from .panel import Panel
+from .roof import Roof
 from .wall import Wall
 
 
@@ -14,7 +16,9 @@ __all__ = [
     "DimensionalElement",
     "Element",
     "ElementType",
+    "Floor",
     "OrientedElement",
     "Panel",
+    "Roof",
     "Wall",
 ]