+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `ElementType` enumeration to be used with `Element.type` property.
 - Added initial `Beam`, `Panel`, and `Wall` classes.
 - Added support for materials and layers through the `Material` and `LayerStack` classes.
+- Added `Floor` and `Roof` classes.
 
 ### Changed